moved action and loader function to their own .ts file
When they were being exported in the .tsx file, they were preventing HMR
