+++ packages/client/src/main.tsx
@@ -5,10 +5,8 @@ import "./index.css";
 import Root from "./routes/root.tsx";
 import ErrorPage from "./error-page.tsx";
 
-import Test, {
-  loader as testLoader,
-  action as testAction,
-} from "./routes/test.tsx";
+import Test from "./routes/test.tsx";
+import { loader as testLoader, action as testAction } from "./routes/test.ts";
 
 import Email from "./routes/emailsTemplate.tsx";
