+++ packages/client/src/routes/test.tsx
@@ -1,55 +1,6 @@
 import { Form, useFetcher, useLoaderData } from "react-router-dom";
-import type { ActionFunction, LoaderFunction } from "react-router";
 import DeleteBinIcon from "../icons/DeleteBinIcon";
-
-declare interface TestLoaderData {
-  data: { id: number; name: string; created_at: string }[];
-}
-
-export const loader: LoaderFunction = async function loader() {
-  try {
-    const response = await fetch(`${__API_PATH__}/test`, {
-      method: "GET",
-    });
-    const data = (await response.json()) as TestLoaderData["data"];
-    return { data };
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    return { data: [] };
-  }
-};
-
-export const action: ActionFunction = async function action({
-  request,
-  params,
-}) {
-  const method = request.method.toLowerCase();
-  switch (method) {
-    case "delete": {
-      const deleteRes = await fetch(`${__API_PATH__}/test/${params.id}`, {
-        method: "DELETE",
-      });
-      if (!deleteRes.ok) throw deleteRes;
-      return { ok: true };
-    }
-    case "post": {
-      const formData = await request.formData();
-      const data = Object.fromEntries(formData);
-
-      const postRes = await fetch(`${__API_PATH__}/test`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(data),
-      });
-      if (!postRes.ok) throw postRes;
-      return { ok: true };
-    }
-    default:
-      throw new Error(`Unsupported method: ${method}`);
-  }
-};
+import type { TestLoaderData } from "./test";
 
 export default function Contact() {
   const { data } = useLoaderData() as TestLoaderData;
@@ -63,7 +14,7 @@ export default function Contact() {
   };
 
   return (
-    <div className="flex min-h-screen flex-col items-center justify-center bg-gray-100">
+    <div className="bg-cus flex min-h-screen flex-col items-center justify-center bg-gray-200">
       <h1 className="mb-8 text-4xl font-bold text-blue-600">Test App!</h1>
       <div className="rounded-lg bg-white p-6 shadow-md">
         <Form method="post" className="mt-4">
